+++ src/main/java/me/drex/instantfeedback/block/SulfurFireBlock.java
@@ -1,0 +1,44 @@
+package me.drex.instantfeedback.block;
+
+import com.blackgear.vanillabackport.common.registries.ModBlocks;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.BaseFireBlock;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class SulfurFireBlock extends BaseFireBlock {
+
+    public SulfurFireBlock(BlockBehaviour.Properties properties) {
+        super(properties, 2.0F);
+    }
+
+    @Override
+    public BlockState updateShape(
+            BlockState state,
+            Direction direction,
+            BlockState neighborState,
+            LevelAccessor level,
+            BlockPos pos,
+            BlockPos neighborPos
+    ) {
+        return this.canSurvive(state, level, pos) ? this.defaultBlockState() : Blocks.AIR.defaultBlockState();
+    }
+
+    @Override
+    public boolean canSurvive(BlockState state, LevelReader level, BlockPos pos) {
+        return canSurviveOnBlock(level.getBlockState(pos.below()));
+    }
+
+    public static boolean canSurviveOnBlock(BlockState state) {
+        return state.is(ModBlocks.SULFUR.get()) || state.is(ModBlocks.POTENT_SULFUR.get());
+    }
+
+    @Override
+    protected boolean canBurn(BlockState state) {
+        return false;
+    }
+}