+++ src/main/java/me/drex/instantfeedback/datagen/ModAdvancementProvider.java
@@ -10,7 +10,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.Level;
-import com.blackgear.vanillabackport.common.registries.ModEntities;
+import com.blackgear.vanillabackport.common.registries.ModEntityTypes;
 import com.blackgear.vanillabackport.common.registries.ModItems;
 
 import java.util.concurrent.CompletableFuture;
@@ -46,7 +46,7 @@ public class ModAdvancementProvider extends FabricAdvancementProvider {
                                 EntityPredicate.wrap(
                                         EntityPredicate.Builder.entity()
                                                 .vehicle(EntityPredicate.Builder.entity()
-                                                        .of(ModEntities.HAPPY_GHAST.get())
+                                                        .of(ModEntityTypes.HAPPY_GHAST)
                                                         .located(LocationPredicate.Builder.location()
                                                                 .setDimension(Level.END)
                                                                 .build())