+++ src/main/java/me/drex/instantfeedback/mixin/CreakingMixin.java
@@ -2,19 +2,23 @@ package me.drex.instantfeedback.mixin;
 
 import me.drex.instantfeedback.config.ConfigManager;
 import com.blackgear.vanillabackport.common.level.entities.creaking.Creaking;
+import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.*;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Creaking.class)
 public abstract class CreakingMixin {
 
-    @ModifyConstant(
+    @Inject(
             method = "createAttributes",
-            constant = @Constant(doubleValue = 0.4),
-            require = 0
+            at = @At("RETURN")
     )
-    private static double increaseMovementSpeed(double original) {
-        return ConfigManager.config().theGardenAwakensBuffCreaking ? 0.45 : original;
+    private static void increaseMovementSpeed(CallbackInfoReturnable<AttributeSupplier.Builder> cir) {
+        if (ConfigManager.config().theGardenAwakensBuffCreaking) {
+            cir.getReturnValue().add(Attributes.MOVEMENT_SPEED, 0.45);
+        }
     }
 
     @Inject(