+++ src/main/java/me/drex/instantfeedback/mixin/sulfur_fire/BlockEntityTypeAccessor.java
@@ -1,0 +1,19 @@
+package me.drex.instantfeedback.mixin.sulfur_fire;
+
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.Set;
+
+@Mixin(BlockEntityType.class)
+public interface BlockEntityTypeAccessor {
+    @Accessor("validBlocks")
+    Set<Block> getValidBlocks();
+
+    @Accessor("validBlocks")
+    @Mutable
+    void setValidBlocks(Set<Block> validBlocks);
+}